// src/app/models/v_PatientsFull.model.ts
export class ViewPatientsFull{
    idPatient: number;
    idTypeDocument: number;
    typeDocument: string;
    documentPatient: string;
    fullName: string;
    idGender: number;
    gender: string;
    dateBirth: Date;
    idCivilStatus: number;
    civilStatus: string;
    addressPatient: string;
    neighborhoodPatient: string;
    codePostal: string;
    idCity: number;
    nameCity: string;
    economicLevel: string;
    telephonePatient: string;
    cellphonePatient: string;
    emailPatient: string;
    occupationPatient: string;
    idInsuranceCompany: number;
    nameInsuranceCompany: string;
    idTypeConnection: number;
    typeConnection: string;
    chaperone: string;
    telephoneChaperone: string;
    inCharge: string;
    idTypeDocumenInCharge: number;
    typeDocumentInCharge: string;
    documentInCharge: string;
    idRelationShipInCharge: number;
    relationShipInCharge: string;
    telephoneInCharge: string;
    residentialZone: string;
    idTypeHearAbout: number;
    typeHearAbout: string;
    typePerson: string;
    dateTimeNew: Date;
    picturePatient: string;
    dataChart: string;
}